prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	long int n = 612852475143;
-	int i;
+	long int i;
 
 	for (i = 2; i <= n; i++)
 		for (; n % i == 0; n /= i)
